ENH: rename internal module _commons.py to lib.py to mark that it's now public
