+++ idefix_cli/_commands/run.py
@@ -15,16 +15,16 @@ from typing import Final
 import inifix
 from rich.prompt import Confirm
 
-from idefix_cli._commons import files_from_patterns
-from idefix_cli._commons import print_err
-from idefix_cli._commons import print_subcommand
-from idefix_cli._commons import print_warning
-from idefix_cli._commons import requires_idefix
+from idefix_cli.lib import files_from_patterns
+from idefix_cli.lib import print_err
+from idefix_cli.lib import print_subcommand
+from idefix_cli.lib import print_warning
+from idefix_cli.lib import requires_idefix
 
 if sys.version_info >= (3, 11):
     from contextlib import chdir
 else:
-    from idefix_cli._commons import chdir
+    from idefix_cli.lib import chdir
 
 # known end messages in Idefix
 KNOWN_SUCCESS: Final = (
