+++ tests/conftest.py
@@ -17,6 +17,6 @@ def inifile(request):
 def isolated_conf_dir(tmp_path, monkeypatch):
     conf_dir = tmp_path / ".config"
     os.makedirs(conf_dir)
-    monkeypatch.setattr("idefix_cli._commons.XDG_CONFIG_HOME", str(conf_dir))
+    monkeypatch.setattr("idefix_cli.lib.XDG_CONFIG_HOME", str(conf_dir))
     monkeypatch.chdir(tmp_path)
     return conf_dir
