+++ tests/test_app_structure.py
@@ -7,9 +7,9 @@ from textwrap import dedent
 
 import pytest
 
-from idefix_cli._commons import print_err
 from idefix_cli._main import _setup_commands
 from idefix_cli._main import main
+from idefix_cli.lib import print_err
 
 
 @pytest.mark.skipif(
